requirements.txt
django
django-conneg
#django-stalefields
-e git+git://github.com/alexsdutton/django-stalefields.git#egg=django_stalefields
-e git+git://github.com/RDFLib/rdflib-jsonld.git#egg=rdflib_jsonld
gdal
jsonfield
jsonpatch
jsonpointer
jsonschema
Markdown
Pillow
psycopg2
python-dateutil
pytz
rdflib
redis
ujson

